requirements.txt
Django==3.0.8
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.2
openpyxl==3.0.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
whitenoise==5.2.0
django-rest-knox==4.1.0

